+++ tests/test_full_example_duckdb.py
@@ -198,10 +198,6 @@ def test_link_only(input, source_l, source_r):
             id="DuckDB link from pandas df",
         ),
         pytest.param(
-            duckdb.read_csv("./tests/datasets/fake_1000_from_splink_demos.csv"),
-            id="DuckDB link from duckdbpyrelation",
-        ),
-        pytest.param(
             pa.Table.from_pandas(
                 pd.read_csv("./tests/datasets/fake_1000_from_splink_demos.csv")
             ),
@@ -323,4 +319,19 @@ def test_small_example_duckdb(tmp_path):
     blocking_rule = "l.dob = r.dob"
     linker.training.estimate_parameters_using_expectation_maximisation(blocking_rule)
 
+    linker.inference.predict()
+
+
+@mark_with_dialects_including("duckdb")
+def test_duckdb_input_is_duckdbpyrelation():
+    df1 = duckdb.read_csv("./tests/datasets/fake_1000_from_splink_demos.csv")
+    df2 = pd.read_csv("./tests/datasets/fake_1000_from_splink_demos.csv")
+
+    settings = SettingsCreator(
+        link_type="link_and_dedupe",
+        comparisons=[cl.ExactMatch("first_name")],
+        blocking_rules_to_generate_predictions=[block_on("first_name", "surname")],
+    )
+    db_api = DuckDBAPI()
+    linker = Linker([df1, df2], settings, db_api)
     linker.inference.predict()